requirements.txt
Django==1.7.4
djangorestframework==3.0
django-debug-toolbar==1.2.2
django-cors-headers==1.0.0